Add log and state change support to contractService
