+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/domain/ContractResultServiceImpl.java
@@ -22,17 +22,21 @@ package com.hedera.mirror.importer.domain;
 
 import com.hederahashgraph.api.proto.java.ContractFunctionResult;
 import com.hederahashgraph.api.proto.java.ContractID;
+import com.hederahashgraph.api.proto.java.ContractLoginfo;
 import java.util.ArrayList;
 import java.util.List;
 import javax.inject.Named;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
+import com.hedera.mirror.common.domain.contract.ContractLog;
 import com.hedera.mirror.common.domain.contract.ContractResult;
+import com.hedera.mirror.common.domain.contract.ContractStateChange;
 import com.hedera.mirror.common.domain.entity.EntityId;
 import com.hedera.mirror.common.domain.transaction.RecordItem;
 import com.hedera.mirror.common.util.DomainUtils;
 import com.hedera.mirror.importer.parser.record.entity.EntityProperties;
+import com.hedera.mirror.importer.util.Utility;
 
 @Log4j2
 @Named
@@ -63,6 +67,62 @@ public class ContractResultServiceImpl implements ContractResultService {
         }
 
         return contractResult;
+    }
+
+    @Override
+    public List<ContractLog> getContractLogs(ContractFunctionResult functionResult, ContractResult contractResult) {
+
+        List<ContractLog> contractLogs = new ArrayList<>();
+        for (int index = 0; index < functionResult.getLogInfoCount(); ++index) {
+            ContractLoginfo contractLoginfo = functionResult.getLogInfo(index);
+
+            ContractLog contractLog = new ContractLog();
+            contractLog.setBloom(DomainUtils.toBytes(contractLoginfo.getBloom()));
+            contractLog.setConsensusTimestamp(contractResult.getConsensusTimestamp());
+            contractLog.setContractId(entityIdService.lookup(contractLoginfo.getContractID()));
+            contractLog.setData(DomainUtils.toBytes(contractLoginfo.getData()));
+            contractLog.setIndex(index);
+            contractLog.setRootContractId(contractResult.getContractId());
+            contractLog.setPayerAccountId(contractResult.getPayerAccountId());
+            contractLog.setTopic0(Utility.getTopic(contractLoginfo, 0));
+            contractLog.setTopic1(Utility.getTopic(contractLoginfo, 1));
+            contractLog.setTopic2(Utility.getTopic(contractLoginfo, 2));
+            contractLog.setTopic3(Utility.getTopic(contractLoginfo, 3));
+            contractLogs.add(contractLog);
+        }
+
+        return contractLogs;
+    }
+
+    @Override
+    public List<ContractStateChange> getContractStateChanges(ContractFunctionResult functionResult,
+                                                             ContractResult contractResult) {
+
+        List<ContractStateChange> contractStateChanges = new ArrayList<>();
+        for (int stateIndex = 0; stateIndex < functionResult.getStateChangesCount(); ++stateIndex) {
+            var contractStateChangeInfo = functionResult.getStateChanges(stateIndex);
+
+            var contractId = entityIdService.lookup(contractStateChangeInfo.getContractID());
+            for (var storageChange : contractStateChangeInfo.getStorageChangesList()) {
+                ContractStateChange contractStateChange = new ContractStateChange();
+                contractStateChange.setConsensusTimestamp(contractResult.getConsensusTimestamp());
+                contractStateChange.setContractId(contractId);
+                contractStateChange.setPayerAccountId(contractResult.getPayerAccountId());
+                contractStateChange.setSlot(DomainUtils.toBytes(storageChange.getSlot()));
+                contractStateChange.setValueRead(DomainUtils.toBytes(storageChange.getValueRead()));
+
+                // If a value of zero is written the valueWritten will be present but the inner value will be
+                // absent. If a value was read and not written this value will not be present.
+                if (storageChange.hasValueWritten()) {
+                    contractStateChange
+                            .setValueWritten(DomainUtils.toBytes(storageChange.getValueWritten().getValue()));
+                }
+
+                contractStateChanges.add(contractStateChange);
+            }
+        }
+
+        return contractStateChanges;
     }
 
     private List<Long> getCreatedContractIds(ContractFunctionResult functionResult) {
